Задачи "Java HashMap: Считаем прогулы". Исправления: логику по работе со студентами вынес в отдельный класс, создал отдельный класс для ключа в мапе, изменил проверку в методе получения студентов, заменил некоторые аннотации на data.
